+++ code/framework/types/String.hpp
@@ -14,19 +14,23 @@ class String : public Data {
 		typedef const SMARTPOINTER<String> c_ptr;
 
 	public:
+		// default constructor
+		String() {};
 		// constructor
 		String(std::string s) : str_(s) {};
+		// copy constructor
+		String(const String& s) : str_(s.str_) {};
 		// destructor
 		~String(void){};
 
 		// returns the value of the string
-		std::string getContent();
+		std::string getContent() const;
 
 		// sets the value of the string
 		void setContent(std::string);
 
 		// returns the data type of this data object: in this case: STRING
-		Type getType() override;
+		Type getType() const override;
 
 	private:
 		// value of the string