+++ CHANGELOG.md
@@ -4,7 +4,10 @@ CHANGELOG of the Unified Image Processing Framework
 Currently under development
 ---------------------------
 
+- Bug: It was not possible to compile and use external modules
+- Bug: Relative paths did not work in a configuration file, when it was called from another directory
 - Enhancement: Added a command to allow listing of all loaded modules `uipf --list`
+- Enhancement: Added Default and Copy Constructors to the Data types
 
 
 Version 1.0.0-beta, July 11, 2015